Reiterating comments and replies
hade new helper function that refreshes post, comments and repiles to get updated counters for reactions, comments and replies. Also ensures proper sorting of comments.
